fix tests

+++ USPEX/IO/OutputRepresentation.py
@@ -69,7 +69,8 @@ class OutputRepresentation(object):
         else:
             raise RuntimeError('Unknown optimizer type in output initialization.')
         self.RES_FOLDER.mkdir(parents=True, exist_ok=True)
-        shutil.copyfile(Path.cwd()/'input.yaml', self.RES_FOLDER/self.PARAMETERS_FILENAME)
+        if (Path.cwd()/'input.yaml').exists():
+            shutil.copyfile(Path.cwd()/'input.yaml', self.RES_FOLDER/self.PARAMETERS_FILENAME)
 
     def presentSystems(self, optimizer):
         if self.targetRepresentation is not None:
